@@ -10,10 +10,9 @@
   "homepage": "http://zpao.github.io/qrcode.react",
   "main": "lib/index.js",
   "scripts": {
-    "flow": "flow",
     "lint": "eslint .",
     "pretty": "prettier --write '{*,.*}.{js,json}' '**/*.{js,json}'",
-    "prepublishOnly": "flow && make clean && make all",
+    "prepublishOnly": "tsc && make clean && make all",
     "prepublish-docs": "make clean && make all",
     "publish-docs": "gh-pages --dist=examples --src='{index.html,bundle.js}'",
     "test": "jest",
@@ -40,28 +39,32 @@
   "devDependencies": {
     "@babel/cli": "^7.13.14",
     "@babel/core": "^7.13.15",
-    "@babel/eslint-parser": "^7.16.5",
     "@babel/plugin-proposal-class-properties": "^7.13.0",
     "@babel/plugin-proposal-object-rest-spread": "^7.13.8",
     "@babel/plugin-proposal-optional-chaining": "^7.13.12",
     "@babel/preset-env": "^7.13.15",
-    "@babel/preset-flow": "^7.13.13",
     "@babel/preset-react": "^7.13.13",
+    "@babel/preset-typescript": "^7.16.7",
+    "@types/jest": "^27.4.0",
+    "@types/node": "^17.0.5",
+    "@types/react": "^17.0.38",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-test-renderer": "^17.0.1",
+    "@typescript-eslint/eslint-plugin": "^5.8.1",
+    "@typescript-eslint/parser": "^5.8.1",
     "babel-loader": "^8.2.2",
-    "eslint": "^8.5.0",
+    "eslint": "^8.6.0",
     "eslint-config-prettier": "^8.1.0",
-    "eslint-plugin-flowtype": "^8.0.3",
     "eslint-plugin-jest": "^25.3.2",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.23.2",
-    "flow-bin": "^0.111.3",
     "gh-pages": "^3.2.3",
     "jest": "^27.4.5",
     "prettier": "^2.2.1",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
     "react-test-renderer": "^17.0.1",
-    "typescript": "^4.4.2",
+    "typescript": "^4.5.4",
     "webpack": "^4.46.0",
     "webpack-cli": "^3.3.12"
   },
