@@ -1,22 +1,15 @@
 module.exports = {
-  parser: '@babel/eslint-parser',
-  parserOptions: {
-    ecmaVersion: 2016,
-    sourceType: 'module',
-    ecmaFeatures: {
-      jsx: true,
-    },
-  },
+  parser: '@typescript-eslint/parser',
   env: {
     es6: true,
     node: true,
     browser: true,
   },
-  plugins: ['react'],
+  plugins: ['react', '@typescript-eslint'],
   extends: [
     'eslint:recommended',
     'plugin:react/recommended',
-    'plugin:flowtype/recommended',
+    'plugin:@typescript-eslint/recommended',
     'plugin:prettier/recommended',
     'prettier',
   ],
@@ -26,7 +19,18 @@ module.exports = {
     },
   },
   rules: {
-    'no-unused-vars': ['error', {ignoreRestSiblings: true}],
+    // Will use the TS version
+    // 'no-unused-vars': ['error', {ignoreRestSiblings: true}],
+    // Meh - don't agree with it
     'react/no-unescaped-entities': 'off',
+    // Meh - I don't feel the need for all things to be const
+    'prefer-const': 'off',
+    // Not as helpful as it seems. I wa
+    // '@typescript-eslint/no-unused-vars': 0,
+    '@typescript-eslint/no-unused-vars': ['error', {ignoreRestSiblings: true}],
+    '@typescript-eslint/no-inferrable-types': [
+      'error',
+      {ignoreParameters: true},
+    ],
   },
 };
